fix: fix logging, typehints, TCP protocol and add more checks in case of failure

+++ DistributedCalendar/Server.py
@@ -1,112 +1,79 @@
-#!/usr/bin/env python3 
+#!/usr/bin/env python3
 
-from __future__ import annotations
-
-import sys 
-import json 
-import socket 
+import sys
+import json
+import socket
 import select
-import pickle 
-import logging 
-import threading 
-from queue import Queue 
+import pickle
+import logging
+import threading
 from typing import Optional, Tuple, List, Dict
-from Calendar import Calendar
-from PersistantCalendar import PersistantHashTable 
-# TODO: import persistent calendar and general calendar 
+from Calendar import Calendar, Repeats
+from PersistantCalendar import PersistantHashTable
 
-# TODO: update logging to be from the central invocation not called in every sub module. 
+
+# TODO: update logging to be from the central invocation not called in every sub module.
 logging.basicConfig(
     level=logging.INFO,
-    format="%(asctime)s [%(peer_name)s : %(levelname)s] %(message)s",
+    format="[%(levelname)s %(asctime)s %(module)s:%(lineno)d] %(message)s",
     datefmt="%H:%M:%S",
 )
-
-def name_server(
-    stop_event: threading.Event,
-    server_port: int,
-    project_name: str,
-    peer_name: str,
-    log: logging.Logger,
-) -> None:
-    """Periodically register this server with the ND catalog via UDP."""
-    s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-    hostname: str = "catalog.cse.nd.edu"
-    port: int = 9097
-    raw_data: Dict[str, str | int] = {
-        "type": "hashtable",
-        "owner": "lmolina3",
-        "port": server_port,
-        "project": project_name,
-        "peer_name": peer_name,
-    }
-    data = json.dumps(raw_data).encode()
-    data_size = len(data)
-    while not stop_event.is_set():
-        sent_amt = 0
-        while sent_amt < data_size:
-            sent = s.sendto(data[: data_size - sent_amt], (hostname, port))
-            sent_amt += sent
-        log.info(f"Send UDP packet for naming to {hostname}")
-        stop_event.wait(60)
-    s.close()
 
 
 # TODO: update type hints
 class Server:
     BUFFER_SIZE = 2**10
-    MAX_ENTRIES = 100 
-    
+    MAX_ENTRIES = 100
+
     def __init__(self, project_name: str, server_name: str, port: int = 0):
         self.project_name: str = project_name
-        self.server_name: str = server_name 
-        self.port: int = port 
-        self.socket: Optional[socket.socket] = None 
+        self.server_name: str = server_name
+        self.port: int = port
+        self.socket: Optional[socket.socket] = None
         self.client_sockets: dict[int, socket.socket] = {}
         self.client_addresses: dict[int, Tuple[str, int]] = {}
         self.threads: List[threading.Thread] = []
         self.stop: threading.Event = threading.Event()
-        self.epoll: Optional[select.poll] = None
-        log = logging.getLogger(__name__)
-        self.log: logging.Logger = logging.LoggerAdapter(log, {"peer_name": server_name})
+        self.epoll: Optional[select.epoll] = None
+        self.log = logging.getLogger(__name__)
         self.log.setLevel(logging.DEBUG)
 
         self.calendar = Calendar()
         self.persistence = PersistantHashTable()
-        # TODO: add presistence calendar
         # TODO: add logic for elections maybe need attributes (synced peers, queue for seralization??)
-        
+
     def start(self) -> None:
         """Initilize server by binding to a port, spawning daemon for nameserver, and listening for requests"""
         self.socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
         try:
-            # set up server socket 
+            # set up server socket
             self.socket.bind(("", self.port))
             self.socket.listen(10)
             self.port = self.socket.getsockname()[1]
             self.socket.setblocking(False)
             self.log.info(f"{self.server_name} listening on \033[32m{self.port}\033[0m")
-            
+
             # set up name service daemon
             t = threading.Thread(
-                target=name_server,
+                target=self.name_server,
                 args=(
                     self.stop,
                     self.port,
                     self.project_name,
                     self.server_name,
-                    self.log
+                    self.log,
                 ),
-                daemon=True
+                daemon=True,
             )
             t.start()
             self.threads.append(t)
-            
+
             self.epoll = select.epoll()
-            assert self.epoll is not None
+            if self.epoll is None:
+                raise ValueError("Epoll object not initialized")
             self.epoll.register(self.socket.fileno(), select.EPOLLIN)
         except Exception as e:
-            self.log.error(f"start: {e}")
+            self.log.error(f"Could not start the Server: {e}")
             sys.exit(1)
 
     def run(self) -> None:
@@ -116,13 +83,13 @@ class Server:
 
         try:
             while True:
-                #TODO: can add election logic here potentially (e.g handle events then handle election)
+                # TODO: can add election logic here potentially (e.g handle events then handle election)
                 self._handle_events()
         except KeyboardInterrupt:
             self.log.info(f"\n{'-'*50}\nShutting down server")
         finally:
             self._cleanup()
-            
+
     def _cleanup(self) -> None:
         """Shutdown server gracefully"""
         if self.epoll is not None:
@@ -136,12 +103,18 @@ class Server:
         self._close_server_socket()
         self.stop.set()
         for t in self.threads:
-            t.join() 
-    
-    def _handle_events(self, timeout: int=1) -> None:
-        """Handle events returned by epoll"""
+            t.join()
+
+    def _handle_events(self, timeout: int = 1) -> None:
+        """Handle events for each client that reaches out to the server"""
+        if self.epoll is None:
+            raise ValueError("Epoll object not initialized")
+        if self.socket is None:
+            raise ValueError("Socket is not initialized")
+
         events = self.epoll.poll(timeout)
         for fileno, event in events:
+            # new connections from client
             if fileno == self.socket.fileno():
                 clt_socket, clt_addr = self.socket.accept()
                 clt_socket.setblocking(False)
@@ -150,11 +123,13 @@ class Server:
                 self.client_sockets[clt_fileno] = clt_socket
                 self.client_addresses[clt_fileno] = clt_addr
                 self.log.info(f"Connection from {clt_addr}")
+            # broken connection from client
             elif event & select.EPOLLHUP:
                 self.log.info(
                     f"closing socket from {self.client_addresses.get(fileno, 'Unknown')}"
                 )
                 self._unregister_socket(fileno)
+            # Receiving data from client
             elif event & select.EPOLLIN:
                 try:
                     request = self._recv_all(fileno)
@@ -162,6 +137,8 @@ class Server:
                         self._unregister_socket(fileno)
                         continue
                     if request == b"":
+                        continue
+                    if not isinstance(request, dict):
                         continue
                     response = self._parse_request(request, fileno)
                     self._send_ack(response, fileno)
@@ -172,17 +149,20 @@ class Server:
                     )
                     self._unregister_socket(fileno)
 
-    
     def _unregister_socket(self, fileno: int) -> None:
+        """Remove connected file descriptors from the interest list, ensuring the kernel stops monitoring it for events"""
+        if self.epoll is None:
+            return
         try:
             self.epoll.unregister(fileno)
         except Exception:
             pass
         self._close_client_socket(fileno)
         self.client_addresses.pop(fileno, None)
-        self.client_sockets.pop(fileno, None) 
+        self.client_sockets.pop(fileno, None)
 
-    def _recv_all(self, fileno: int)-> dict[str, str] | bytes | None:
+    def _recv_all(self, fileno: int) -> dict[str, str] | bytes | None:
+        """Recieve entire payload from file descriptor of interest, return the payload if not malformed or broken connection"""
         if fileno not in self.client_sockets:
             self.log.error(f"recv_all: {fileno} not in client sockets")
             return {}
@@ -190,7 +170,7 @@ class Server:
         client_address = self.client_addresses[fileno]
         header = b""
 
-        # Recieve Header 
+        # Recieve Header
         while b"\n" not in header:
             try:
                 data = client_socket.recv(self.BUFFER_SIZE)
@@ -208,7 +188,7 @@ class Server:
         buffer = header[delim_idx + 1 :]
         read_amt = len(buffer)
 
-        # Recieve body 
+        # Recieve body
         while read_amt < data_size:
             try:
                 data = client_socket.recv(self.BUFFER_SIZE)
@@ -229,8 +209,9 @@ class Server:
                 f"recv_all: Deserialization failed from {client_address}: {e}"
             )
             return None
-        
+
     def _send_ack(self, payload: dict[str, str], fileno: int) -> None:
+        """Send acknowledgment of the request to file descriptor"""
         if fileno not in self.client_sockets:
             self.log.error(f"send_ack: {fileno} not in client sockets")
             return
@@ -243,8 +224,9 @@ class Server:
         finally:
             client_socket.setblocking(False)
 
-
-    def _parse_request(self, request: dict[str, str], fileno: int) -> dict[str, str]:
+    def _parse_request(
+        self, request: dict[str, str], fileno: int
+    ) -> dict[str, str | int]:
         if fileno not in self.client_sockets:
             self.log.error(f"parse_request: {fileno} not in client sockets")
             return {"status": "failure", "error": f"{fileno} not in client sockets"}
@@ -259,31 +241,28 @@ class Server:
             }
         try:
             method = request.get("method", "")
+            params = request.get("params", {})
             match method:
                 case "create":
-                    name = request.get("name")
-                    start = request.get("start")
-                    end = request.get("end")
-                    description = request.get("description")
-                    location = request.get("location")
-                    repeats = request.get("repeats")
-                    ident = self.persistence.create(name, start, end, description, location, repeats) 
+                    valid, msg = self._validate_rpc(method, params)
+                    if not valid:
+                        return {"method": "create", "status": "failure", "error": msg}
+                    ident = self.persistence.create(**params)
                     if ident is None:
                         return {"method": "create", "status": "failure"}
                     return {"method": "create", "status": "success", "ident": ident}
                 case "delete":
+                    valid, msg = self._validate_rpc(method, params)
+                    if not valid:
+                        return {"method": "create", "status": "failure", "error": msg}
                     ident = request.get("ident")
-                    self.persistence.delete(ident)
+                    self.persistence.delete(**params)
                     return {"method": "delete", "status": "success"}
                 case "modify":
-                    ident = request.get("ident")
-                    name = request.get("name")
-                    start = request.get("start")
-                    end = request.get("end")
-                    description = request.get("description")
-                    location = request.get("location")
-                    repeats = request.get("repeats")
-                    ident = self.persistence.modify(ident, name, start, end, description, location, repeats) 
+                    valid, msg = self._validate_rpc(method, params)
+                    if not valid:
+                        return {"method": "create", "status": "failure", "error": msg}
+                    ident = self.persistence.modify(**params)
                     if ident is None:
                         return {"method": "modify", "status": "failure"}
                     return {"method": "modify", "status": "success", "ident": ident}
@@ -301,7 +280,29 @@ class Server:
             self.log.error(f"{e}")
             return {"status": "failure", "error": str(e)}
 
+    def _validate_rpc(
+        self, method: str, params: dict[str, str | int | Repeats | None]
+    ) -> Tuple[bool, str]:
+        if not params:
+            return (
+                False,
+                f"{method} parameters empty, look at API for specific paramters",
+            )
+        if not isinstance(params, dict):
+            return False, f"Parameters must be passed in as a dictionary"
+        if (method == "delete" or method == "modify") and params.get("ident", "") == "":
+            return False, f"{method} requires the parameter ident"
+        if method == "create" or method == "modify":
+            if params.get("name", None) is None:
+                return False, f"{method} requires the parameter name"
+            if params.get("start", None) is None:
+                return False, f"{method} requires the parameter start"
+            if params.get("end", None) is None:
+                return False, f"{method} requires the parameter end"
+        return True, ""
+
     def _close_client_socket(self, fileno: int) -> None:
+        """Close file descriptor socket gracefully"""
         if fileno not in self.client_sockets:
             self.log.error(f"close_client_socket: {fileno} not found in client sockets")
             return
@@ -312,8 +313,9 @@ class Server:
             except OSError:
                 pass
             client_socket.close()
-    
+
     def _close_server_socket(self) -> None:
+        """Close the server socket gracefully"""
         if self.socket:
             try:
                 self.socket.shutdown(socket.SHUT_RDWR)
@@ -322,7 +324,36 @@ class Server:
             self.socket.close()
             self.socket = None
 
-            
+    def name_server(
+        self,
+        stop_event: threading.Event,
+        server_port: int,
+        project_name: str,
+        peer_name: str,
+        log: logging.Logger,
+    ) -> None:
+        """Periodically register this server with the ND catalog via UDP."""
+        s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+        hostname: str = "catalog.cse.nd.edu"
+        port: int = 9097
+        raw_data: Dict[str, str | int] = {
+            "type": "hashtable",
+            "owner": "lmolina3",
+            "port": server_port,
+            "project": project_name,
+            "peer_name": peer_name,
+        }
+        data = json.dumps(raw_data).encode()
+        data_size = len(data)
+        while not stop_event.is_set():
+            sent_amt = 0
+            while sent_amt < data_size:
+                sent = s.sendto(data[: data_size - sent_amt], (hostname, port))
+                sent_amt += sent
+            log.info(f"Send UDP packet for naming to {hostname}")
+            stop_event.wait(60)
+        s.close()
+
 
 def main() -> None:
     if len(sys.argv) != 3:
